UI updates

+++ src/app/components/project/location-list/location-details/section-list/section-list.component.ts
@@ -1,5 +1,5 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {MatDialog} from "@angular/material/dialog";
+import {Component, EventEmitter, Input, OnInit, Output, SimpleChange} from '@angular/core';
+import {MatDialog, MatDialogConfig} from "@angular/material/dialog";
 import {BuildingLocation, Section} from "../../../../../common/models/buildingLocation";
 import {
   OrchestratorCommunicationService
@@ -10,6 +10,7 @@ import {Store} from "@ngrx/store";
 import {ProjectState, SectionState} from "../../../../../app-state-service/store/project-state-model";
 import {OrchestratorEventName} from "../../../../../orchestrator-service/models/orchestrator-event-name";
 import { environment } from '../../../../../../environments/environment';
+import { DeleteConfirmationModalComponent } from '../../../../../forms/delete-confirmation-modal/delete-confirmation-modal.component';
 
 @Component({
   selector: 'app-section-list',
@@ -23,6 +24,7 @@ export class SectionListComponent implements OnInit{
   @Output() sectionsDeletionComplete = new EventEmitter<boolean>();
   location_!: BuildingLocation;  // @Input() location!: BuildingLocation
   sections: Section[] = [];
+  @Input() isLoading!: boolean;
   @Input()
   set location(location: BuildingLocation) {
     this.location_ = location;
@@ -55,6 +57,17 @@ export class SectionListComponent implements OnInit{
     });
   }
 
+ngOnChanges(changes: { [property: string]: SimpleChange }) {
+    // Extract changes to the input property by its name
+    let change: SimpleChange = changes['isLoading']; 
+    if(change?.currentValue){
+      
+    }
+    // Whenever the data in the parent changes, this method gets triggered
+    // You can act on the changes here. You will have both the previous
+    // value and the  current value here.
+}
+
   private getSections(location: BuildingLocation) {
     if (this.projectState === ProjectState.INVASIVE) {
       // TODO: Check Logic for Invasive
@@ -75,17 +88,29 @@ export class SectionListComponent implements OnInit{
     }
 
   deleteElement($event: string) {
-    let id = $event;
-    let url = `${environment.apiURL}/section/${id}`;
-    this.httpsRequestService.deleteHttpData(url).subscribe(
-      (response: any) => {
-        console.log(response);
-        this.sectionsDeletionComplete.emit(true);
-      }
-      , error => {
-        console.log(error);
-      }
-    );
 
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.width = "450px";
+    // dialogConfig.height = "140px";
+    const dialogRef = this.dialog.open(DeleteConfirmationModalComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe(data => {
+      if(data.confirmed){
+
+        let id = $event;
+        let url = `${environment.apiURL}/section/${id}`;
+        this.httpsRequestService.deleteHttpData(url).subscribe(
+          (response: any) => {
+            console.log(response);
+            this.sectionsDeletionComplete.emit(true);
+          }
+          , error => {
+            console.log(error);
+            this.sectionsDeletionComplete.emit(false);
+          }
+        );
+
+      }})
   }
 }
